+++ spec/transpec/configuration_spec.rb
@@ -9,19 +9,20 @@ module Transpec
 
     context 'by default' do
       [
-        [:convert_should?,            true],
-        [:convert_oneliner?,          true],
-        [:convert_should_receive?,    true],
-        [:convert_stub?,              true],
-        [:convert_have_items?,        true],
-        [:convert_its?,               true],
-        [:convert_deprecated_method?, true],
-        [:parenthesize_matcher_arg?,  true],
-        [:forced?,                    false],
-        [:skip_dynamic_analysis?,     false],
-        [:negative_form_of_to,        'not_to'],
-        [:boolean_matcher_type,       :conditional],
-        [:form_of_be_falsey,          'be_falsey']
+        [:convert_should?,                                        true],
+        [:convert_oneliner?,                                      true],
+        [:convert_should_receive?,                                true],
+        [:convert_stub?,                                          true],
+        [:convert_have_items?,                                    true],
+        [:convert_its?,                                           true],
+        [:convert_deprecated_method?,                             true],
+        [:parenthesize_matcher_arg?,                              true],
+        [:add_receiver_arg_to_any_instance_implementation_block?, true],
+        [:forced?,                                                false],
+        [:skip_dynamic_analysis?,                                 false],
+        [:negative_form_of_to,                                    'not_to'],
+        [:boolean_matcher_type,                                   :conditional],
+        [:form_of_be_falsey,                                      'be_falsey']
       ].each do |attribute, value|
         describe "##{attribute}" do
           subject { configuration.send(attribute) }